+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2026-07-07
+
 ### Fixed
 - `:windows` transport Windows-account detection. The transport locates its
   script dir under `C:\Users\<user>\AppData\Local\Temp`, but the user was